+++ src/leoButtons.ts
@@ -77,21 +77,9 @@ export class LeoButtonNode extends vscode.TreeItem {
         private _leoIntegration: LeoIntegration
     ) {
         super(button.name);
-        // Cleanup button name of any directives starting with '@'
-        // super(
-        //     button.name.split(" ")
-        //         .filter(item => item[0] !== "@")
-        //         .join(" ")
-        //         .trim()
-        // );
 
         this._id = utils.getUniqueId();
-        // Setup this instance (just differentiate 'script-button' for now)
-        this.command = {
-            command: Constants.COMMANDS.CLICK_BUTTON,
-            title: '',
-            arguments: [this]
-        };
+
         this._isAdd = (this.button.index.startsWith(Constants.BUTTON_STRINGS.NULL_WIDGET) &&
             this.button.name === Constants.BUTTON_STRINGS.SCRIPT_BUTTON);
         this.rclicks = button.rclicks ? button.rclicks : [];