+++ src/constants.ts
@@ -36,9 +36,13 @@ export class Constants {
     public static JS_FILE_EXTENSION: string = "leojs";
 
     public static LEO_LANGUAGE_PREFIX: string = "leobody."; // all lowercase
+    public static LEO_WRAP_SUFFIX: string = ".wrap"; // all lowercase.
+
     public static URI_LEO_SCHEME: string = "leo";
     public static URI_FILE_SCHEME: string = "file";
     public static URI_SCHEME_HEADER: string = "leo:/";
+    public static URI_UNTITLED_SCHEME: string = 'untitled';
+    public static OUTPUT_CHANNEL_LANGUAGE = 'Log';
     public static FILE_OPEN_FILTER_MESSAGE: string = "Leo Files";
     public static UNTITLED_FILE_NAME: string = "untitled";
     public static RECENT_FILES_KEY: string = "leoRecentFiles";
@@ -70,12 +74,34 @@ export class Constants {
     public static LOG_ALERT_COLOR: string = 'red';
 
     /**
+     * Supported Languages
+     */
+    public static LANGUAGES = [
+        "plain",
+        "python",
+        "javascript",
+        "typescript",
+        "c",
+        "cpp",
+        "css",
+        "fortran",
+        "fortran90",
+        "html",
+        "java",
+        "json",
+        "markdown",
+        "php",
+        "restructuredtext",
+        "rust",
+    ];
+
+    /**
      * * Minimal Leo Editor and Leo server versions
      */
     public static MIN_SERVER_VERSION_NUMBER = {
         major: 1,
         minor: 0,
-        patch: 8
+        patch: 9,
     };
 
     /**
@@ -239,8 +265,8 @@ export class Constants {
         IGNORED: " They were ignored.", // with leading space
         TOO_FAST: "leoInteg is busy! ", // with trailing space
         MINIMUM_LEO_VERSION_STRING: "Please update your Leo Installation: Leo 6.7.5 is recommended.",
-        UNKNOWN_LANGUAGE_NOT_SUPPORTED: "Language not yet supported.",
-        LANGUAGE_NOT_SUPPORTED: " language not yet supported.", // with leading space
+        UNKNOWN_LANGUAGE_NOT_SUPPORTED: "Language coloring not yet supported.",
+        LANGUAGE_NOT_SUPPORTED: " language coloring not yet supported.", // with leading space
         MINIBUFFER_BUTTON_START: "@button-",
         MINIBUFFER_RCLICK_START: "@rclick-",
         MINIBUFFER_SCRIPT_BUTTON: "script-button",
@@ -253,7 +279,17 @@ export class Constants {
         MINIBUFFER_COMMAND: "$(zap) @command",
         MINIBUFFER_BAD_COMMAND: "$(error) Not Available",
         MINIBUFFER_HISTORY_LABEL: "Minibuffer History",
-        MINIBUFFER_HISTORY_DESC: "Choose from last run commands..."
+        MINIBUFFER_HISTORY_DESC: "Choose from last run commands...",
+
+        ZOOM_LEVEL_RANGE_LIMIT: "Value for zoom level should be between -12 and 12",
+        FONT_SIZE_RANGE_LIMIT: "Value for font size should be between 6 and 30",
+        BODY_WRAP_RECOMMEND: "'leojsbody' wordWrap settings are recommended (some currently missing)",
+        BODY_WRAP_SET: "'leojsbody' wordWrap settings were set",
+        FIX_IT: "Fix it",
+        ENABLE_PREVIEW_SET: "'Enable Preview' setting was set",
+        ENABLE_PREVIEW_RECOMMEND: "'Enable Preview' setting is recommended (currently disabled)",
+        CLOSE_EMPTY_CLEARED: "'Close Empty Groups' setting was cleared",
+        CLOSE_EMPTY_RECOMMEND: "'Close Empty Groups' setting is NOT recommended!",
     };
 
     /**
@@ -413,6 +449,7 @@ export class Constants {
     public static CONTEXT_FLAGS = {
         // Main flags for connection and opened file
         STARTUP_FINISHED: "leoStartupFinished", // Initial extension finished auto-server-start-connect
+        LEO_OPENING_FILE: "leoOpeningFile",
         STARTING_SERVER: "leoStartingServer",
         CONNECTING: "leoConnecting", // Initial extension finished auto-server-start-connect
         BRIDGE_READY: "leoBridgeReady", // Connected to leoBridge
@@ -559,6 +596,7 @@ export class Constants {
         GET_VERSION: "!get_version",
         GET_LEOID: "!get_leoid",
         SET_LEOID: "!set_leoid",
+        HANDLE_UNL: "!handle_unl",
         // * Server Commands
         GET_COMMANDS: "!get_all_leo_commands", // "getCommands",
         APPLY_CONFIG: "!set_config", // "applyConfig",
@@ -732,6 +770,7 @@ export class Constants {
         STOP_SERVER: Constants.NAME + ".stopServer",
         CONNECT: Constants.NAME + ".connectToServer",
         SET_LEOID: Constants.NAME + ".setLeoID",
+        HANDLE_UNL: Constants.NAME + ".handleUnl",
         SET_OPENED_FILE: Constants.NAME + ".setOpenedFile",
         OPEN_FILE: Constants.NAME + ".openLeoFile", // sets focus on BODY
         CLEAR_RECENT_FILES: Constants.NAME + ".clearRecentFiles",
@@ -942,6 +981,7 @@ export class Constants {
         TOGGLE_FIND_WORD_OPTION: Constants.NAME + ".toggleFindWordOption",
         TOGGLE_FIND_SEARCH_BODY_OPTION: Constants.NAME + ".toggleFindSearchBodyOption",
         TOGGLE_FIND_SEARCH_HEADLINE_OPTION: Constants.NAME + ".toggleFindSearchHeadlineOption",
+        SET_BODY_WRAP_SETTINGS: Constants.NAME + ".setBodyWrapSettings",
         SET_ENABLE_PREVIEW: Constants.NAME + ".setEnablePreview",
         CLEAR_CLOSE_EMPTY_GROUPS: Constants.NAME + ".clearCloseEmptyGroups",
     };