+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 1.0.19
 
-- Fixed ...
+- Fixed tooltips in find panel
+- Added UNL support: Requires Leo 6.7.7
+- Fixed 'goto-script' refresh cycle.
+- @button panel now run scripts with run icon button instead of selecting the entry.
 
 ## 1.0.18